test: fix tests and move them to cron
