+++ .github/workflows/tests-cron.yml
@@ -1,0 +1,44 @@
+name: Upgrade dependencies
+
+on:
+  workflow_dispatch:
+  schedule:
+    - cron: "0 9 * * 1-5" # At 09:00 AM, Monday through Friday
+
+jobs:
+  tests:
+    name: Long running Tests
+    runs-on: ubuntu-latest
+    continue-on-error: true # we want the PR to popup even if sth goes wrong
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v2
+
+      - name: Use Node.js
+        uses: actions/setup-node@v3
+        with:
+          node-version: 16
+          registry-url: "https://registry.npmjs.org/"
+          scope: "@talend"
+          cache: "yarn"
+
+      - name: Upgrade dependencies
+        run: |
+          yarn --frozen-lockfile --ignore-scripts
+          yarn test:cron
+
+      - name: Create Issue on Failed workflow
+        if: ${{ failure() }}
+        uses: dacbd/create-issue-action@v1
+        with:
+          token: ${{ secrets.GITHUB_TOKEN }}
+          title: Action workflow failed.
+          body: |
+            ### Context
+            [Failed Run](https://github.com/${{ github.repository }}/actions/runs/${{ github.run_id }})
+            [Codebase](https://github.com/${{ github.repository }}/tree/${{ github.sha }})
+            Workflow name - `${{ github.workflow }}`
+            Job -           `${{ github.job }}`
+            status -        `${{ job.status }}`
+          assignees: SomeUsername,AnotherUsername
