+++ fork/module-to-cdn/index.test.js
@@ -1,8 +1,6 @@
-const validate = require('validate-npm-package-name');
 const modules = require('./modules.json');
 const fn = require('.');
 const {parseToSemverIfPossible} = require('./version');
-const {cachedGet, getModuleInfo, getAllVersions, getRangeEdgeVersions} = require('./cache');
 
 /* eslint-disable no-param-reassign */
 /* eslint-disable no-console */
@@ -23,94 +21,6 @@ const t = {
     },
     pass: () => expect(true).toBe(true)
 };
-
-async function testModule(t, moduleName, version, env) {
-    const cdnConfig = fn(moduleName, version, {env});
-
-    await testCdnConfig(t, cdnConfig, moduleName, version);
-}
-
-// https://stackoverflow.com/a/31625466/3052444
-function isValidVarName(name) {
-    try {
-        if (name.includes('.')) {
-            // E.g. ng.core would cause errors otherwise:
-            name = name.split('.').join('_');
-        }
-
-        if (name.includes('-')) {
-            name = `"${name}"`;
-            // eslint-disable-next-line no-eval
-            return eval(`(function() { a = {${name}:1}; a[${name}];}); true`);
-        }
-
-        if (name.includes('}')) {
-            return false;
-        }
-
-        // eslint-disable-next-line no-eval
-        return eval(`(function() { a = {${name}:1}; a.${name}; var ${name}; }); true`);
-    } catch (error) {
-        console.error(error);
-        return false;
-    }
-}
-async function testCdnConfig(t, cdnConfig, moduleName, version) {
-    t.notDeepEqual(cdnConfig, null);
-
-    t.is(cdnConfig.name, fn.getModuleName(moduleName));
-    t.truthy(cdnConfig.url);
-    t.true(cdnConfig.url.includes(version));
-
-    await t.notThrowsAsync(async () => {
-        if (cdnConfig.url.endsWith('/')) {
-            return;
-        }
-
-        let data;
-        try {
-            const response = await cachedGet(cdnConfig.url);
-            data = response.data;
-        } catch (error) {
-            throw new Error(error.message);
-        }
-
-        if (cdnConfig.var) {
-            t.true(isValidVarName(cdnConfig.var));
-
-            const content = data.replace(/ /g, '');
-            t.true(
-                content.includes(`.${cdnConfig.var}=`) ||
-                    content.includes(`["${cdnConfig.var}"]=`) ||
-                    content.includes(`['${cdnConfig.var}']=`) ||
-                    // Immutable 3 is clear, the script is global and just do Immutable =
-                    content.includes(`${cdnConfig.var}=`) ||
-                    content.includes(`function${cdnConfig.var}(`),
-                `${cdnConfig.var} not found in the content ${cdnConfig.url}`
-            );
-        }
-    }, cdnConfig.url);
-}
-
-async function testNextModule(t, moduleName, env) {
-    const tags = getModuleInfo(moduleName)['dist-tags'];
-
-    if (!tags.next) {
-        return t.pass();
-    }
-
-    const nextVersion = tags.next;
-
-    const cdnConfig = fn(moduleName, nextVersion, {env});
-
-    if (!cdnConfig) {
-        return t.pass(`no next support for ${moduleName}`);
-    }
-
-    await testCdnConfig(t, cdnConfig, moduleName, nextVersion);
-}
-
-const moduleNames = Object.keys(modules);
 
 describe('module-to-cdn', () => {
     it('basic', () => {
@@ -202,43 +112,3 @@ describe('module-to-cdn', () => {
         });
     });
 });
-
-function limit(m) {
-    if (process.env.LIMIT) {
-        return process.env.LIMIT.includes(`;${m};`);
-    }
-
-    return true;
-}
-
-// eslint-disable-next-line no-restricted-syntax
-for (const importPath of moduleNames.filter(m => limit(m))) {
-    const {validForNewPackages} = validate(importPath);
-
-    const prodTest = () => {
-        testNextModule(t, importPath, 'production');
-    };
-    const devTest = () => {
-        testNextModule(t, importPath, 'development');
-    };
-    if (!validForNewPackages) {
-        describe('invalid package name', () => {
-            xit(`prod: ${importPath}@next`, prodTest);
-            xit(`dev: ${importPath}@next`, devTest);
-        });
-    } else {
-        describe('valid package name', () => {
-            it(`prod: ${importPath}@next`, prodTest);
-            it(`dev: ${importPath}@next`, devTest);
-        });
-
-        const moduleName = fn.getModuleName(importPath);
-        const allVersions = getAllVersions(moduleName);
-        const versionRanges = Object.keys(modules[importPath].versions);
-        const testVersions = [].concat(...versionRanges.map(getRangeEdgeVersions(allVersions)));
-        describe.each(testVersions)('prod: $importPath @ $version', version => {
-            testModule(t, moduleName, version, 'production');
-            testModule(t, moduleName, version, 'development');
-        });
-    }
-}
